move scraper domain to root core module
