+++ core/src/main/java/GroceryFamily/GroceryElders/domain/Currency.java
@@ -1,0 +1,3 @@
+package GroceryFamily.GroceryElders.domain;
+
+public enum Currency {EUR}
