--- scraper/core/src/main/java/GroceryFamily/GroceryDad/domain/Currency.java
@@ -1,3 +1,0 @@
-package GroceryFamily.GroceryDad.domain;
-
-public enum Currency {EUR}
